refactor: createNews
